% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarditza_1.R
\name{estandarditza_1}
\alias{estandarditza_1}
\title{estandarditza raw data en format long}
\usage{
estandarditza_1(indicador, grep_edicio, grep_seccio)
}
\arguments{
\item{indicador}{data frame o tibble que conté dadesd'inici, i que ha de contenir l'any de referència de les dades, les seccions censals i com a mínim una variable més amb la variable de referència.}

\item{grep_edicio}{expressió regular per identificar el nom de la variable que designa les edicions}

\item{description}{Estandarditza les dades d'origen en format long i dona noms consistents a les variables. Els arxius de dades es llegeixen fora d'aquesta funció.}

\item{grep_seccions}{expressió regular per identificar el nom de la variable que designa les seccions censals}
}
\value{
tibble en format long
}
\description{
estandarditza raw data en format long
}
\examples{
data(ex_data, package = "interpola")
estandarditza_1(ex_data$indicador, "ANYO", "SECCIO$")
}
