% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Dades de joguina pels exemples 
inclou dos df: un indicador amb la taxa d'atur segons el cens i una submostra de les dades cadastrals del 2022 per l'AMB}
\format{
1 Named list amb dues taules  
\describe{  
  \item{indicador}{taula de dades d'atur}  
  \item{cad_AMB}{taula amb dades cadastre}  
  \item{shapes}{ESRI shapefile d'exemple (crs = 4326)}  
  ...  
}
}
\usage{
ex_data
}
\description{
Dades de joguina pels exemples 
inclou dos df: un indicador amb la taxa d'atur segons el cens i una submostra de les dades cadastrals del 2022 per l'AMB
}
\keyword{datasets}
