% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fes_mapa.R
\name{fes_mapa}
\alias{fes_mapa}
\title{fes_mapa}
\usage{
fes_mapa(shape_data, nom_indicador, ciutat = "Barcelona", focus = 14)
}
\arguments{
\item{shape_data}{conjunt de dades tipus shapefile que contingui les geometries de les divisions territorials d'interès. També cal que contingui la variable que es vol dibuixar. Veure exemple.}

\item{nom_indicador}{cadena de text que representa el nom de la variable que es vol dibuixar.}

\item{ciutat}{cadena de text amb el nom de la ciutat on es posa el focus inicial del mapa. Per defecte, Barcelona.}

\item{focus}{Xifra que indica el nivell de zoom. Per defecte, 14.}
}
\value{
Un widget html amb un choropleth
}
\description{
Fes mapa (choropleth) amb els resultats de les interpolacions. Per defecte, fa servir `colorNumeric` per definir els colors.
}
\examples{
data("ex_data", package = "interpola") # carrega dades de mostra
sp <- ex_data$shapes \%>\% # afegeix variable pels colors a dades shape
mutate(
 Tatur = ex_data$indicador$taxa_atur[match(ex_data$shapes$Sc2021, ex_data$indicador$sc_2021)],
 Tatur = round(readr::parse_double(Tatur),2)
)
fes_mapa(shape_data = sp, nom_indicador = "Tatur", ciutat = "Badalona", focus = 14) # fes gràfic
}
