% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifica_seccionat_2.R
\name{identifica_seccionat_2}
\alias{identifica_seccionat_2}
\title{Identifica seccionat de les dades d'origen}
\usage{
identifica_seccionat_2(indicador)
}
\arguments{
\item{indicador}{conjunt de dades que ha passat per la fase 1}
}
\value{
cadena de text amb l'any de l'edició que té el màxim número de matches, i missatge de text amb el número de matches
}
\description{
Identifica seccionat de les dades d'origen
}
\examples{
data(ex_data, package = "interpola")
identifica_seccionat_2(ex_data$indicador)
}
